kernel minor changes

+++ Kernel/include/pipes.h
@@ -5,11 +5,12 @@
 #define IN_USE 1
 #define EMPTY 0
 
-typedef struct t_pipe {
+typedef struct t_pipe
+{
   int id;
   char buffer[PIPE_BUFFER_SIZE];
   int writeIndex, readIndex;
-  long totalProcesses;
+  int totalProcesses;
   int writeLock, readLock;
   int state;
 } t_pipe;
